+++ src/Pages/Dashboard/Filters/Government/AllGovFilters.js
@@ -11,7 +11,7 @@ import DeleteItem from "../../../../Components/DeleteItem/DeleteItem.js";
 import { useNavigate } from 'react-router-dom';
 
 export default function AllGovFilters() {
-  
+  const role = localStorage.getItem("role")
   const navigate = useNavigate();
   const [loadId, setLoadId] = useState(false);
   const [filters, setFilters] = useState([]);
@@ -103,13 +103,13 @@ export default function AllGovFilters() {
                         تعديل
                       </Button>
                     </td>
-                    <DeleteItem
+                    {role==='admin'&&<DeleteItem
                       id={selectedItemId}
                       setId={setSelectedItemId}
                       itemId={item.id}
                       DeleteFun={handleDelete}
                       load={loadId}
-                    />
+                    />}
                   </tr>
                 ))}
               </tbody>
